% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_gene_list.R
\name{compare_gene_list}
\alias{compare_gene_list}
\title{compare_gene_list}
\usage{
compare_gene_list(gene_sets)
}
\arguments{
\item{gene_sets}{A named list of gene vectors. Each element should be a character vector of gene names.}
}
\value{
A ggplot2 Venn diagram object is printed to the console.
}
\description{
Visualize overlaps among gene lists using a Venn diagram.
}
