% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mcl.R
\name{find_mcl}
\alias{find_mcl}
\title{find_mcl}
\usage{
find_mcl(
  network,
  min_node = 10,
  expansion = 2,
  inflation = 1.5,
  plot_network = FALSE,
  label_genes = TRUE
)
}
\arguments{
\item{network}{An igraph object representing the input network.}

\item{min_node}{Integer. Minimum number of nodes to retain a cluster (default = 10).}

\item{expansion}{Integer. MCL expansion parameter (default = 2).}

\item{inflation}{Numeric. MCL inflation parameter (default = 1.5).}

\item{plot_network}{Logical. Whether to plot the resulting MCL subnetworks (default = FALSE).}

\item{label_genes}{Logical. Whether to display vertex labels in the plot (default = TRUE).}
}
\value{
A named list of igraph objects representing the largest filtered MCL subnetworks.
}
\description{
Detect MCL (Markov Cluster Algorithm) subnetworks in a graph and optionally plot the results.
}
