% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pagerank.R
\name{get_pagerank}
\alias{get_pagerank}
\title{get_pagerank}
\usage{
get_pagerank(network)
}
\arguments{
\item{network}{An igraph object representing a gene or protein interaction network.}
}
\value{
A named numeric vector of PageRank scores for each node. A ggnet2 plot is displayed.
}
\description{
Compute PageRank scores for a network and visualize using ggnet2.
}
