% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_directed_graph.R
\name{make_directed_graph}
\alias{make_directed_graph}
\title{Create Directed Graph with Omnipath Directionality}
\usage{
make_directed_graph(network, node_size = 5, label_genes = TRUE, label_size = 2)
}
\arguments{
\item{network}{igraph object representing the initial network}

\item{node_size}{Numeric, size of the network nodes}

\item{label_genes}{Logical, whether to label genes (nodes) or not}

\item{label_size}{Numeric, font size for node labels}
}
\value{
A directed igraph network visualized using ggnet2
}
\description{
This function takes an igraph network, retrieves directionality from Omnipath, adjusts edge directions accordingly, and plots a directed network.
}
\examples{
make_directed_graph(network = your_network_object)

}
