% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ora.R
\name{run_ora}
\alias{run_ora}
\title{run_ora}
\usage{
run_ora(
  genes,
  ontology = "BP",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.05,
  showCategory = 10,
  plot = FALSE
)
}
\arguments{
\item{genes}{A character vector of gene symbols}

\item{ontology}{Character. Ontology to use for GO analysis ("BP", "MF", or "CC")}

\item{pvalueCutoff}{Numeric. P-value cutoff for enrichment (default = 0.05)}

\item{qvalueCutoff}{Numeric. Q-value (adjusted p-value) cutoff for enrichment (default = 0.05)}

\item{showCategory}{Integer. Number of top enriched terms to show in the plot (default = 10)}

\item{plot}{Logical. Whether to display a cnetplot of enriched terms (default = FALSE)}
}
\value{
A data frame of enriched GO terms. Returns NULL if fewer than 5 genes are mapped or no terms are enriched.
}
\description{
Perform Over-Representation Analysis (ORA) using Gene Ontology (GO) and optionally plot cnetplot
}
